Add update notification
